+++ app/(tabs)/_layout.tsx
@@ -45,7 +45,7 @@ export default function TabLayout() {
       <Tabs.Screen
         name="explore"
         options={{
-          title: 'Character',
+          title: 'Explore',
           tabBarIcon: ({ color }) => <MaterialCommunityIcons name="face-man-profile" size={24} color="black" />,
         }}
       />
